Some changes, removed requestty, give error message if cant find platforms, improved stressor functions becuase some where being optimized out. Still gotta fix somethings

+++ src/reporting.rs
@@ -15,7 +15,7 @@ pub struct BackgroundReport {
 }
 
 pub fn watch_in_background(
-    stop_temperature: Option<i64>,
+    stop_temperature: Option<u8>,
     duration: Option<Duration>,
     system: &mut System,
     start_time: Instant,
